# feat: implement Phase 3 medium-priority vulnerability detectors achieving 100% coverage