… crates/detectors/src/reward_calculation.rs
… crates/detectors/src/timestamp.rs
… crates/detectors/src/weak_commit_reveal.rs